+++ Ratio.h
@@ -4,13 +4,13 @@
 class Ratio {
 private:
 	float sum;
-	size_t amount;
+	float amount;
 public:
 	Ratio();
 
 	void addResult(const Result& result);
 	void addRatio(Ratio& ratio);
-	void addRatio(const float sum_, const size_t amount_);
+	void addRatio(const float sum_, const float amount_);
 	float getChance();
-	size_t getAmount();
+	float getAmount();
 };
